# perf(ext/http): faster req_url string assembly